Introduce Notification, NotificationProducer
There is a circular dependency between SlackNotifier and ActiveNotfier.
An ActiveNotfier is created by the SlackNotifier, which is responsible
for constructing the notification for Slack and calling back to the
SlackNotifier for sending the message.

This change models the Notification as a class and makes the ActiveNotifier
implement the NotificationProducer interface. This will allow the
SlackNotifier to delegate to the ActiveNotifier for message production,
but ultimately leave the SlackNotifier responsible for sending the
notification.